requirements.txt
python-chess
torch
numpy